Fixing routes for remaining areas of the app
